refactor: many improvements in UniswapV2LikeFetcher
