+++ packages/oracle-node/src/fetchers/evm-chain/ethereum/sushiswap-on-chain/all-sushiswap-fetchers.ts
