+++ packages/oracle-node/src/fetchers/evm-chain/ethereum/sushiswap-on-chain/sushiswap-fetchers-config.json
