+++ packages/oracle-node/src/fetchers/uniswap-v2-like/UniswapV2.abi.json
